+++ lizard_graph/urls.py
@@ -8,6 +8,7 @@ from django.views.generic import TemplateView
 from lizard_ui.urls import debugmode_urlpatterns
 
 from lizard_graph.views import GraphView
+from lizard_graph.views import graph_window
 
 
 admin.autodiscover()
@@ -17,6 +18,9 @@ urlpatterns = patterns(
     url(r'^$',
         GraphView.as_view(),
         name="lizard_graph_graph_view"),
+    url(r'^window/$',
+        graph_window,
+        name="lizard_graph_graph_window"),
     url(r'^examples/$',
         TemplateView.as_view(template_name="lizard_graph/examples.html")),
     )